create a superclass of StatTracker  named  FileOpener
