+++ lib/file_opener.rb
@@ -1,0 +1,17 @@
+class FileOpener
+  def self.open_files(locations)
+    games = CSV.open locations[:games], headers: true, header_converters: :symbol
+    teams = CSV.open locations[:teams], headers: true, header_converters: :symbol
+    game_teams = CSV.open locations[:game_teams], headers: true, header_converters: :symbol
+    make_rows([games, teams, game_teams])
+  end
+
+  def self.make_rows(array)
+    dummy_array = array.map do |file|
+      file.map do |row|
+        row
+      end
+    end
+    StatTracker.new(dummy_array[0], dummy_array[1], dummy_array[2])
+  end
+end
